requirements.txt
aiogram>=3.22.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0